+++ src/Common/AzurePSCmdlet.cs
@@ -1,4 +1,4 @@
-// ----------------------------------------------------------------------------------
+﻿// ----------------------------------------------------------------------------------
 //
 // Copyright Microsoft Corporation
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -21,6 +21,11 @@ using System.Reflection;
 using Microsoft.Azure.Common.Authentication;
 using Microsoft.Azure.Common.Authentication.Models;
 using Microsoft.WindowsAzure.Commands.Common;
+using Newtonsoft.Json;
+using System.IO;
+using System.Management.Automation.Host;
+using System.Text;
+using System.Threading;
 
 namespace Microsoft.WindowsAzure.Commands.Utilities.Common
 {
@@ -34,6 +39,20 @@ namespace Microsoft.WindowsAzure.Commands.Utilities.Common
         private RecordingTracingInterceptor _httpTracingInterceptor;
 
         private DebugStreamTraceListener _adalListener;
+        protected static AzurePSDataCollectionProfile _dataCollectionProfile = null;
+        protected static string _errorRecordFolderPath = null;
+        protected const string _fileTimeStampSuffixFormat = "yyyy-MM-dd-THH-mm-ss-fff";
+
+        protected AzurePSQoSEvent QosEvent;
+
+        protected virtual bool IsUsageMetricEnabled {
+            get { return false; }
+        }
+
+        protected virtual bool IsErrorMetricEnabled
+        {
+            get { return true; }
+        }
 
         /// <summary>
         /// Gets the PowerShell module name used for user agent header.
@@ -60,10 +79,123 @@ namespace Microsoft.WindowsAzure.Commands.Utilities.Common
         }
 
         /// <summary>
-        /// Cmdlet begin process. Write to logs, setup Http Tracing and initialize profile and adds user agent.
+        /// Initialize the data collection profile
+        /// </summary>
+        protected static void InitializeDataCollectionProfile()
+        {
+            if (_dataCollectionProfile != null && _dataCollectionProfile.EnableAzureDataCollection.HasValue)
+            {
+                return;
+            }
+
+            // Get the value of the environment variable for Azure PS data collection setting.
+            string value = Environment.GetEnvironmentVariable(AzurePSDataCollectionProfile.EnvironmentVariableName);
+            if (!string.IsNullOrWhiteSpace(value))
+            {
+                if (string.Equals(value, bool.FalseString, StringComparison.OrdinalIgnoreCase))
+                {
+                    // Disable data collection only if it is explictly set to 'false'.
+                    _dataCollectionProfile = new AzurePSDataCollectionProfile(true);
+                }
+                else if (string.Equals(value, bool.TrueString, StringComparison.OrdinalIgnoreCase))
+                {
+                    // Enable data collection only if it is explictly set to 'true'.
+                    _dataCollectionProfile = new AzurePSDataCollectionProfile(false);
+                }
+            }
+
+            // If the environment value is null or empty, or not correctly set, try to read the setting from default file location.
+            if (_dataCollectionProfile == null)
+            {
+                string fileFullPath = Path.Combine(AzureSession.ProfileDirectory, AzurePSDataCollectionProfile.DefaultFileName);
+                if (File.Exists(fileFullPath))
+                {
+                    string contents = File.ReadAllText(fileFullPath);
+                    _dataCollectionProfile = JsonConvert.DeserializeObject<AzurePSDataCollectionProfile>(contents);
+                }
+            }
+
+            // If the environment variable or file content is not set, create a new profile object.
+            if (_dataCollectionProfile == null)
+            {
+                _dataCollectionProfile = new AzurePSDataCollectionProfile();
+            }
+        }
+
+        /// <summary>
+        /// Get the data collection profile
+        /// </summary>
+        protected static AzurePSDataCollectionProfile GetDataCollectionProfile()
+        {
+            if (_dataCollectionProfile == null)
+            {
+                InitializeDataCollectionProfile();
+            }
+
+            return _dataCollectionProfile;
+        }
+
+        /// <summary>
+        /// Check whether the data collection is opted in from user
+        /// </summary>
+        /// <returns>true if allowed</returns>
+        public static bool IsDataCollectionAllowed()
+        {
+            if (_dataCollectionProfile != null &&
+                _dataCollectionProfile.EnableAzureDataCollection.HasValue &&
+                _dataCollectionProfile.EnableAzureDataCollection.Value)
+            {
+                return true;
+            }
+
+            return false;
+        }
+
+        /// <summary>
+        /// Save the current data collection profile Json data into the default file path
+        /// </summary>
+        /// <param name="profile"></param>
+        protected abstract void SaveDataCollectionProfile();
+
+        protected bool CheckIfInteractive()
+        {
+            bool interactive = true;
+            if (this.Host == null || this.Host.UI == null || this.Host.UI.RawUI == null)
+            {
+                interactive = false;
+            }
+            else
+            {
+                try
+                {
+                    var test = this.Host.UI.RawUI.KeyAvailable;
+                }
+                catch
+                {
+                    interactive = false;
+                }
+            }
+
+            if (!interactive && !_dataCollectionProfile.EnableAzureDataCollection.HasValue)
+            {
+                _dataCollectionProfile.EnableAzureDataCollection = false;
+            }
+            return interactive;
+        }
+
+        /// <summary>
+        /// Prompt for the current data collection profile
+        /// </summary>
+        /// <param name="profile"></param>
+        protected abstract void PromptForDataCollectionProfileIfNotExists();
+
+        /// <summary>
+        /// Cmdlet begin process. Write to logs, setup Http Tracing and initialize profile
         /// </summary>
         protected override void BeginProcessing()
         {
+            PromptForDataCollectionProfileIfNotExists();
+            InitializeQosEvent();
             if (string.IsNullOrEmpty(ParameterSetName))
             {
                 WriteDebugWithTimestamp(string.Format("{0} begin processing without ParameterSet.", this.GetType().Name));
@@ -95,6 +227,7 @@ namespace Microsoft.WindowsAzure.Commands.Utilities.Common
         /// </summary>
         protected override void EndProcessing()
         {
+            LogQosEvent();
             string message = string.Format("{0} end processing.", this.GetType().Name);
             WriteDebugWithTimestamp(message);
 
@@ -124,7 +257,14 @@ namespace Microsoft.WindowsAzure.Commands.Utilities.Common
 
         protected new void WriteError(ErrorRecord errorRecord)
         {
-            FlushDebugMessages();
+            FlushDebugMessages(IsDataCollectionAllowed());
+            if (QosEvent != null && errorRecord != null)
+            {
+                QosEvent.Exception = errorRecord.Exception;
+                QosEvent.IsSuccess = false;
+                LogQosEvent(true);
+            }
+            
             base.WriteError(errorRecord);
         }
 
@@ -225,12 +365,93 @@ namespace Microsoft.WindowsAzure.Commands.Utilities.Common
             WriteObject(customObject);
         }
 
-        private void FlushDebugMessages()
+        private void FlushDebugMessages(bool record = false)
         {
+            if (record)
+            {
+                RecordDebugMessages();
+            }
+
             string message;
             while (_debugMessages.TryDequeue(out message))
             {
                 base.WriteDebug(message);
+            }
+        }
+
+        protected abstract void InitializeQosEvent();
+
+        private void RecordDebugMessages()
+        {
+            // Create 'ErrorRecords' folder under profile directory, if not exists
+            if (string.IsNullOrEmpty(_errorRecordFolderPath) || !Directory.Exists(_errorRecordFolderPath))
+            {
+                _errorRecordFolderPath = Path.Combine(AzureSession.ProfileDirectory, "ErrorRecords");
+                Directory.CreateDirectory(_errorRecordFolderPath);
+            }
+
+            CommandInfo cmd = this.MyInvocation.MyCommand;
+
+            string filePrefix = cmd.Name;
+            string timeSampSuffix = DateTime.Now.ToString(_fileTimeStampSuffixFormat);
+            string fileName = filePrefix + "_" + timeSampSuffix + ".log";
+            string filePath = Path.Combine(_errorRecordFolderPath, fileName);
+
+            StringBuilder sb = new StringBuilder();
+            sb.Append("Module : ").AppendLine(cmd.ModuleName);
+            sb.Append("Cmdlet : ").AppendLine(cmd.Name);
+
+            sb.AppendLine("Parameters");
+            foreach (var item in this.MyInvocation.BoundParameters)
+            {
+                sb.Append(" -").Append(item.Key).Append(" : ");
+                sb.AppendLine(item.Value == null ? "null" : item.Value.ToString());
+            }
+
+            sb.AppendLine();
+
+            foreach (var content in _debugMessages)
+            {
+                sb.AppendLine(content);
+            }
+
+            AzureSession.DataStore.WriteFile(filePath, sb.ToString());
+        }
+
+        /// <summary>
+        /// Invoke this method when the cmdlet is completed or terminated.
+        /// </summary>
+        protected void LogQosEvent(bool waitForMetricSending = false)
+        {
+            if (QosEvent == null)
+            {
+                return;
+            }
+
+            QosEvent.FinishQosEvent();
+
+            if (!IsUsageMetricEnabled && (!IsErrorMetricEnabled || QosEvent.IsSuccess))
+            {
+                return;
+            }
+
+            if (!IsDataCollectionAllowed())
+            {
+                return;
+            }
+
+            WriteDebug(QosEvent.ToString());
+
+            try
+            {
+                MetricHelper.LogQoSEvent(QosEvent, IsUsageMetricEnabled, IsErrorMetricEnabled);
+                MetricHelper.FlushMetric(waitForMetricSending);
+                WriteDebug("Finish sending metric.");
+            }
+            catch (Exception e)
+            {
+                //Swallow error from Application Insights event collection.
+                WriteWarning(e.ToString());
             }
         }
 
@@ -244,10 +465,19 @@ namespace Microsoft.WindowsAzure.Commands.Utilities.Common
         /// <param name="action">The action code</param>
         protected void ConfirmAction(bool force, string actionMessage, string processMessage, string target, Action action)
         {
+            if (QosEvent != null)
+            {
+                QosEvent.PauseQoSTimer();
+            }
+            
             if (force || ShouldContinue(actionMessage, ""))
             {
                 if (ShouldProcess(target, processMessage))
-                {
+                {                 
+                    if (QosEvent != null)
+                    {
+                        QosEvent.ResumeQosTimer();
+                    }
                     action();
                 }
             }
@@ -255,7 +485,10 @@ namespace Microsoft.WindowsAzure.Commands.Utilities.Common
 
         protected virtual void Dispose(bool disposing)
         {
-            _adalListener.Dispose();
+            if (_adalListener != null)
+            {
+                _adalListener.Dispose();
+            }
         }
 
         public void Dispose()