+++ src/ResourceManager/Resources/Commands.Resources/ResourceGroupDeployments/GetAzureResourceGroupDeploymentCommand.cs
@@ -47,11 +47,10 @@ namespace Microsoft.Azure.Commands.Resources
                     new List<string>() { ProvisioningState }
             };
 
-            if(!string.IsNullOrEmpty(Name))
+            if(!string.IsNullOrEmpty(ProvisioningState))
             {
-                WriteWarning("The parameter 'Name' in Get-AzureResourceGroupDeployment cmdlet is being renamed to DeploymentName and will be updated in a future release.");
+                WriteWarning("The ProvisioningState parameter is being deprecated and will be removed in a future release.");
             }
-
             WriteObject(ResourcesClient.FilterResourceGroupDeployments(options), true);
         }
     }