+++ src/ResourceManager/Sql/Commands.Sql/Index Recommendations/Cmdlet/GetAzureSqlDatabaseIndexRecommendations.cs
@@ -23,7 +23,7 @@ using Microsoft.Azure.Common.Authentication.Models;
 namespace Microsoft.Azure.Commands.Sql.Cmdlet
 {
     /// <summary>
-    /// Defines the Get-AzureSqlDatabaseIndexRecommendations cmdlet
+    /// Defines the Get-AzureRMSqlDatabaseIndexRecommendations cmdlet
     /// </summary>
     [Cmdlet(VerbsCommon.Get, "AzureRMSqlDatabaseIndexRecommendations", ConfirmImpact = ConfirmImpact.None)]
     public class GetAzureSqlDatabaseIndexRecommendations : AzureSqlCmdletBase<IEnumerable<IndexRecommendation>, AzureSqlDatabaseIndexRecommendationAdapter>
@@ -105,7 +105,7 @@ namespace Microsoft.Azure.Commands.Sql.Cmdlet
         /// <returns>The recommended index adapter</returns>
         protected override AzureSqlDatabaseIndexRecommendationAdapter InitModelAdapter(AzureSubscription subscription)
         {
-            return new AzureSqlDatabaseIndexRecommendationAdapter(DefaultProfile.DefaultContext);
+            return new AzureSqlDatabaseIndexRecommendationAdapter(DefaultProfile.Context);
         }
     }
 }