+++ src/ServiceManagement/HDInsight/Commands.HDInsight/Cmdlet/StopAzureHDInsightJobCmdlet.cs
@@ -123,6 +123,7 @@ namespace Microsoft.WindowsAzure.Management.HDInsight.Cmdlet.PSCmdlets
         /// <inheritdoc />
         protected override void EndProcessing()
         {
+            this.WriteWarning(AzureHdInsightPowerShellConstants.AsmWarning);
             try
             {
                 this.command.Logger = this.Logger;