+++ src/ServiceManagement/Services/Commands.Test/Websites/GetAzureWebsiteLogTests.cs
@@ -86,7 +86,7 @@ namespace Microsoft.WindowsAzure.Commands.Test.Websites
                     }
                 }
             };
-            Cache.AddSite(currentProfile.DefaultContext.Subscription.Id.ToString(), website);
+            Cache.AddSite(currentProfile.Context.Subscription.Id.ToString(), website);
             websitesClientMock.Setup(c => c.GetWebsite(websiteName, slot))
                 .Returns(website);
         }