+++ ChangeLog.md
@@ -1,4 +1,6 @@
 ﻿## 2015.09.03 version 0.9.8
+* Azure Redis Cache cmdlets 
+  * New-AzureRMRedisCache - 'RedisVersion' parameter is deprecated.
 * Azure Compute (ARM) Cmdlets
   * Added -Launch parameter for Get-AzureRemoteDesktopFile cmdlet
   * Added Id parameter for VM cmdlets to support piping scenario without ResourceGroupName parameter
@@ -7,7 +9,6 @@
   * Changed the output format of Get image cmdlets as a table
   * Fixed Set-AzureVMAccessExtension cmdlet
 * Azure Compute (Service Management) cmdlets
-  * Changed the warning message to a non-terminating error message for ResourceNotFound in VM cmdlets
   * Exposed ComputeImageConfig in Get-AzurePlatformVMImage cmdlet
   * Fixed Publish-AzurePlatformExtension and Set-AzurePlatformExtension cmdlets
 * Azure Backup - added the following cmdlets
@@ -44,6 +45,12 @@
   * Get-AzureSqlServerActiveDirectoryAdministrator
   * Set-AzureSqlServerActiveDirectoryAdministrator
   * Remove-AzureSqlServerActiveDirectoryAdministrator
+* SQL Server VM cmdlets (ARM)
+  * New-AzureVMSqlServerAutoPatchingConfig
+  * New-AzureVMSqlServerAutoBackupConfig
+  * Set-AzureVMSqlServerExtension
+  * Get-AzureVMSqlServerExtension
+  * Remove-AzureVMSqlServerExtension
 	
 ## 2015.08.17 version 0.9.7
 * Azure Profile cmdlets